requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
psycopg2-binary
gunicorn
Pillow